bugfix: call lookupClassMethod instead
This compiler tool would check whether `Observer` have method `observerWithTarget:action:` or not. But the post said "This method starts by looking for message expressions that have Observer as the receiver, and observerWithTarget:action: as the selector, and then looks at the target and checks if the method exists. " Is there something wrong?